refactoring following core / autoconfigure / starter structure
