+++ ci-droid-autoconfigure/src/main/java/com/societegenerale/cidroid/config/CiDroidAutoConfiguration.java
@@ -1,4 +1,4 @@
-package com.societegenerale.cidroid;
+package com.societegenerale.cidroid.config;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationFeature;
@@ -13,7 +13,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 
 @Configuration
-public class CiDroidConfiguration {
+public class CiDroidAutoConfiguration {
 
     @Bean
     public MappingJackson2HttpMessageConverter customJackson2HttpMessageConverter() {
