+++ ci-droid-starter/pom.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.societegenerale.ci-droid</groupId>
+    <artifactId>ci-droid-starter</artifactId>
+    <version>1.0.1-SNAPSHOT</version>
+    <packaging>jar</packaging>
+
+    <name>CI droid starter</name>
+
+    <parent>
+        <groupId>com.societegenerale.ci-droid</groupId>
+        <artifactId>ci-droid-parent</artifactId>
+        <version>1.0.1-SNAPSHOT</version>
+    </parent>
+
+
+    <description>CI-droid REST API autoconfigure</description>
+    <url>https://github.com/societe-generale/ci-droid</url>
+
+    <developers>
+        <developer>
+            <id>vincent-fuchs</id>
+            <name>Vincent FUCHS</name>
+            <email>vincent.fuchs@gmail.com</email>
+        </developer>
+    </developers>
+
+    <licenses>
+        <license>
+            <name>Apache 2</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <java.version>1.8</java.version>
+    </properties>
+
+    <issueManagement>
+        <url>https://github.com/societe-generale/ci-droid/issues</url>
+        <system>GitHub</system>
+    </issueManagement>
+
+    <!-- required by maven-release-plugin -->
+    <scm>
+        <url>https://github.com/societe-generale/ci-droid</url>
+        <connection>scm:git:git@github.com:societe-generale/ci-droid.git</connection>
+        <developerConnection>scm:git:git@github.com:societe-generale/ci-droid.git</developerConnection>
+        <tag>HEAD</tag>
+    </scm>
+
+    <distributionManagement>
+        <snapshotRepository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+        </snapshotRepository>
+        <repository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/service/local/staging/deploy/maven2/</url>
+        </repository>
+    </distributionManagement>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.societegenerale.ci-droid</groupId>
+            <artifactId>ci-droid-autoconfigure</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+
+    </dependencies>
+
+</project>
